pyproject.toml

[project]
name = "voice-to-cart"
version = "0.1.0"
description = "Backend FastAPI para procesar mensajes (texto/voz), interpretar intención con LLM y agregar artículos a un carrito."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "PyYAML>=6.0.1",
    "httpx>=0.27.0",
    "pytest>=8.2.0"
]

[tool.pytest.ini_options]
addopts = "-q"
